requirements.txt
flask>=2.0.2,<3.0
google-cloud-storage>=1.42.0,<2.0
tenacity>=7.0.0,<8.0
